fix: fix -Wextra warnings across kernel_types, ttnn/ (r23 batch)
- kernel_types.hpp: add = {} defaults to DataMovementConfig and ComputeConfig
  container fields (compile_args, defines, named_compile_args, etc.)
- program_descriptors.hpp: add = {} defaults to CBDescriptor format_descriptors
- matmul_program_config_types.hpp: add = std::nullopt to fused_activation fields
- tensor_impl.cpp: fix int64_t vs size_t in strides.size() comparison
- tensor/xtensor/partition.cpp: cast dim to size_t vs tensor_dims
- distributed_tensor.cpp: cast cluster_axis.value() to size_t
- reports.cpp: use uint32_t for page_index loop
- compute_throttle_utils.cpp: cast num_cores to uint32_t vs constexpr uint32_t
- reshape_common.cpp: use size_t for index loop
- op_profiler_json.cpp: cast json .value() to size_t vs packed_size

+++ ttnn/core/distributed/distributed_tensor.cpp
@@ -568,7 +568,7 @@ std::unique_ptr<TensorToMesh> shard_tensor_to_mesh_mapper(
                 .mesh_shape_override = MeshShape(mesh_device.num_devices())}));
     }
     TT_FATAL(
-        cluster_axis.value() >= 0 && cluster_axis.value() < mesh_device.shape().dims(),
+        cluster_axis.value() >= 0 && static_cast<size_t>(cluster_axis.value()) < mesh_device.shape().dims(),
         "Cluster axis {} is out of range for mesh device with {} dimensions",
         cluster_axis.value(),
         mesh_device.shape().dims());
